feat(trading): teams information for referral page (#5643)
