+++ apps/trading/lib/hooks/Team.graphql
