+++ libs/i18n/src/locales/en/trading.json
@@ -60,6 +60,8 @@
   "Create a referral code": "Create a referral code",
   "Create": "Create",
   "Create a team": "Create a team",
+  "Create a simple referral code to enjoy the referrer commission outlined in the current referral program": "Create a simple referral code to enjoy the referrer commission outlined in the current referral program",
+  "Make your referral code a Team to compete in Competitions with your friends, appear in leaderboards on the <0>Competitions Homepage</0>, and earn rewards": "Make your referral code a Team to compete in Competitions with your friends, appear in leaderboards on the <0>Competitions Homepage</0>, and earn rewards",
   "Current tier": "Current tier",
   "DISCLAIMER_P1": "Vega is a decentralised peer-to-peer protocol that can be used to trade derivatives with cryptoassets. The Vega Protocol is an implementation layer (layer one) protocol made of free, public, open-source or source-available software. Use of the Vega Protocol involves various risks, including but not limited to, losses while digital assets are supplied to the Vega Protocol and losses due to the fluctuation of prices of assets.",
   "DISCLAIMER_P2": "Before using the Vega Protocol, review the relevant documentation at docs.vega.xyz to make sure that you understand how it works. Conduct your own due diligence and consult your financial advisor before making any investment decisions.",
@@ -120,6 +122,7 @@
   "Get started": "Get started",
   "Give Feedback": "Give Feedback",
   "Go back and try again": "Go back and try again",
+  "Got to competitions": "Go to competitions",
   "Governance": "Governance",
   "Governance vote for this market has been rejected": "Governance vote for this market has been rejected",
   "Governance vote for this market is valid and has been accepted": "Governance vote for this market is valid and has been accepted",
